fix(ui): split streamed pull-request refresh from chat events
Keep the merged event owner below the unchanged 700-line lint bound.
Move existing link parsing and bounded stream refresh without changing
behavior, reply visibility, terminal recovery, or metadata invalidation.

The merged guard fails at 710 lines before and passes after. Existing
117 UI tests, forced UI types, lint, build, and isolated P0 review pass.
